backend/requirements.txt
# Project Prism - Python Dependencies

# Django
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
dj-database-url>=2.1,<3.0

# Database
psycopg2-binary>=2.9,<3.0

# Server
gunicorn>=21.0,<22.0

# Security
cryptography>=41.0,<42.0

# Utilities
python-dotenv>=1.0,<2.0

# Development
black>=23.0,<24.0
flake8>=6.0,<7.0
pytest-django>=4.5,<5.0
